refactor: organize pages by category
