+++ apps/front/src/pages/Resources/Podcast/index.tsx
@@ -8,11 +8,11 @@ import { Link } from 'react-router-dom';
 
 import { trpc } from '@sovereign-academy/api-client';
 
-import readingRabbit from '../../assets/resources/reading-rabbit.svg';
-import { Button } from '../../atoms/Button';
-import { Card } from '../../atoms/Card';
-import { ResourceLayout } from '../../components';
-import { useRequiredParams } from '../../utils';
+import readingRabbit from '../../../assets/resources/reading-rabbit.svg';
+import { Button } from '../../../atoms/Button';
+import { Card } from '../../../atoms/Card';
+import { ResourceLayout } from '../../../components';
+import { useRequiredParams } from '../../../utils';
 
 const { useGreater } = BreakPointHooks(breakpointsTailwind);
