+++ apps/front/src/pages/index.ts
@@ -1,1 +1,23 @@
-export * from './Home';
+// About
+export { Manifesto } from './About/Manifesto';
+export { SponsorsAndContributors } from './About/SponsorsAndContributors';
+// Courses
+export { CourseChapter } from './Courses/CourseChapter';
+export { CourseDetails } from './Courses/CourseDetails';
+export { CoursesExplorer } from './Courses/CoursesExplorer';
+// Home
+export { Home } from './Home';
+// Misc
+export { NotFound } from './Misc/NotFound';
+export { UnderConstruction } from './Misc/UnderConstruction';
+// Resources
+export { Book } from './Resources/Book';
+export { Books } from './Resources/Books';
+export { Builder } from './Resources/Builder';
+export { Builders } from './Resources/Builders';
+export { Podcast } from './Resources/Podcast';
+export { Podcasts } from './Resources/Podcasts';
+// Tutorials
+export { Tutorial } from './Tutorials/Tutorial';
+export { TutorialCategory } from './Tutorials/TutorialCategory';
+export { Tutorials } from './Tutorials';
